Move GalleryItem to its own file
